+++ __tests__/server.tests/api_v1.test.js
@@ -1,15 +1,18 @@
 import moxios from 'moxios'
 import request from 'supertest'
 
-import initApp from '../../__mocks__/mockapp'
-import { v1 } from '../../src/server/api/index'
+import initServer from '../../__mocks__/mockapp'
+import { v1 } from '../../src/server/routes/index'
 
 const BASE_URL = '/api/v1'
 const URL = `${BASE_URL}/tests`
 
 describe('API v1 tests', () => {
+  let server
+
   beforeEach(() => {
     moxios.install()
+    server = initServer(v1, BASE_URL)
   })
 
   afterEach(() => {
@@ -17,54 +20,42 @@ describe('API v1 tests', () => {
   })
 
   test('GET /api/v1', async(done) => {
-    const app = initApp(v1, BASE_URL)
-
-    const res = await request(app).get(BASE_URL)
+    const res = await request(server).get(BASE_URL)
 
     expect(res.body.message).toEqual('Hello /api/v1!')
     done()
   })
 
   test('GET /api/v1/tests', async(done) => {
-    const app = initApp(v1, BASE_URL)
-
-    const res = await request(app).get(URL)
+    const res = await request(server).get(URL)
 
     expect(res.body.message).toEqual('GET /api/tests')
     done()
   })
 
   test('GET /api/v1/tests/:testId', async(done) => {
-    const app = initApp(v1, BASE_URL)
-
-    const res = await request(app).get(`${URL}/123`)
+    const res = await request(server).get(`${URL}/123`)
 
     expect(res.body.message).toEqual('GET /api/test')
     done()
   })
 
   test('POST /api/v1/tests:testId', async(done) => {
-    const app = initApp(v1, BASE_URL)
-
-    const res = await request(app).post(`${URL}/123`)
+    const res = await request(server).post(`${URL}/123`)
 
     expect(res.body.message).toEqual('POST /api/test')
     done()
   })
 
   test('PUT /api/v1/tests:testId', async(done) => {
-    const app = initApp(v1, BASE_URL, v1)
-
-    const res = await request(app).put(`${URL}/123`)
+    const res = await request(server).put(`${URL}/123`)
 
     expect(res.body.message).toEqual('PUT /api/test')
     done()
   })
 
   test('PUT /api/v1/tests:testId', async(done) => {
-    const app = initApp(v1, BASE_URL)
-
-    const res = await request(app).delete(`${URL}/123`)
+    const res = await request(server).delete(`${URL}/123`)
 
     expect(res.body.message).toEqual('DELETE /api/test')
     done()
